clear the values of Calender before creating a Date.
The set method documentation states that it "sets the values for
the fields: YEAR, MONTH, DAY, HOUR, MINUTE, and SECOND." But
most importantly "Previous values of other fields are retained.
If this is not desired, call clear() first".

This means that TestDateUtil.date() is not idempotent. So calling
it repeatedly with the same values will not be guaranteed to create
equivalent Dates. This is a problem especially when testing where
you expect to check the dates for equality.