sync: keep recursed consumer json out of entitlements
When exporting entitlements, we descend into referenced objects and
export them too. Some we use, like pools. Others we do not, like
consumers.

When a newer candlepin generates an export for an older candlepin, the
newer candlepin's consumer object contains a releaseVer field that
references a object, where the old candlepin expects a string. jackson
is not able to handle this case, so the parser fails.

Since we don't need the consumer object from inside the entitlement,
simply skip serializing it.