834591: 0 or no sockets count as infinite on products
Ensure candlepin, like subscription-manager, treats a product attribute
value of 0 for sockets as infinite sockets, or no value for sockets as
infinite. Do this both for select best pools and compliance.

In other words, if a pool has sockets=0 or no sockets, it will fully
entitle a consumer's product, regardless of how many sockets the
consumer has.