Remove clobbering of SLA level case
We want to treat PREMIUM = premium = Premium when dealing with SLAs, but
forcing the sla values to all be upper case makes them look really ugly
in the gui. Ignore case for comparison, and simply use the first
instance of an SLA's case we see for display; it will probably be the
only instance, anyways.