Improve performance of getting the owner info.
With large numbers of pools, getting the consumer types from the database start to show up as a decent performance hit. This patch stores the types in a local hashmap so that they can be looked up without a DB hit. The only issue should be if the types change during this call. This seems like a row risk.