Reduce amount of data in GET /consumers.
When serializing lists of consumers, we can save ourselves a lot of
serialization time by skipping the identity cert and facts, which are
not likely to be used when listing.

Introduced a @HateoasArrayIgnore annotation to apply to fields that
should be skipped whenever being serialized in the context of an array.