Filter pools with too many content sets from V1 autobinds.
Had hoped to do this in rules but the information is not available, and
exposing it would introduce an API change which would break on all past
candlepin releases when it recieved the new rules in a manifest or the
like. (unless we introduce some form of rules/JS API versioning)

Because the actual exception we throw when a system tries to use such a
pool is already in core java code, I decided to do this in java as well.

The pools with too much content for V1 consumers will still appear in
the list of available pools, they just won't ever be used in an
autobind/dry-run scenario. This is a PM request that the customer we
prompted to upgrade their client if they try to use such a pool.