ownerinfo: replace pool iteration with hql queries
Ownerinfo is used by a number of web front ends to display 'dashboard'
type info. As such, it has to be fast, so we don't slow down that
initial page.

Rather than iterating over each pool that an owner has access to,
which will slow down as the owner gets more pools, use a series of HQL
queries, which are instead bounded by the number of consumer types and
product families. These values should be much smaller, and change less
often.

On my machine, with 100 test consumers and 350 pools, the call takes
0.25s instead of 1.5s.