Fix certv1 content filtering on autobinds.
Mistakenly coded to only check for too much content on marketing
products, which can't have content. Needed to be checking each provided
product.

Because we could now filter down to 0 pools in the select best pools
method, I had to move the exception that was thrown one layer above in
as the client tools expect this behavior when nothing can be
successfully autobound.