Improvements for the performance test script.
With performance work done recently we can now run this much quicker so
bumping up the subscription count to 1000 (very high) and the system
count to 5000. (probably still a little low but I want to limit the setup
time)

Now uses 5 installed products, and the subscriptions each provide some
random combination of those.

Tried threading the registrations but there is no gain, each autobind
attempt arrives and gets the same pools as the preferred best fit, but
then the binds are locked so they all queue up waiting for each other to
finish. Even worse the entitlements quickly exhaust as they're all
trying to get the same pools.

Stacking is still disabled, suspect there is a bug here in behavior I am
seeing but can't pin it down yet.