certv3: store payload and sig in the cert column
key, cert, payload, and signature on entitlement certificate are all PEM
formatted blobs. for payload and signature to be used by the client, it
had to append them onto the cert field before saving to disk (as you
might do with key, if you wanted to have a single file). To reduce the
work the client has to do, and to reduce api changes/schema changes,
just append payload and signature onto the contents of cert before
storing.