Added candlepin.enable_cert_v3 config property.
This property will be used to determine if we should enable/disable
candlepins cert V3 functionality and fall back to old V1 certs.

Allowed spec tests to bypass the enable cert v3 config property.
The V3 spec tests were failing due to defaulting the
candlepin.enable_cert_v3 config property to false. Because there
was no way for a spec test to determine if V3 cert support was
disabled, a custom fact was added to test consumer so that the
config property could be overridden in our tests.

CP will check for that fact (system.testing) and if it exists,
CP will create certs as normal.

This should be removed once cert V3 support has been added elsewhere.