Add new mechanism for detecting/overriding import conflicts.
Added several new types of manifest import conflicts which can be
overridden by the caller. On import 409 responses will also include
progrematic keys which other projects can use to distinguish why the
import failed. The request can be retried with a multi-valued force
parameter to override.

force=true will behave as it did before, ignoring an older manifest
creation date and importing anyhow.

Added distributor conflicts which allow the caller to link their org to
a new upstream distributor. (should they have deleted it accidentally,
etc) This may cause drastic changes to their subscriptions, but not if
used for a new distributor in the same upstream org and using the same
pools.

Signature checking code is brought back but disabled until other
projects are ready for it.

We also now distinguish between an older manifest, and the same manifest
currently imported.