Ignore old rules in database after a Candlepin upgrade.
We were correctly catching incoming old rules during manifest imports
and ignoring them in favor of those included in our rpm. We were not
handling old rules in the database after a candlepin upgrade however,
leading to major failures as we tried to use old rules that didn't work
on the new Candlepin.

Added a column in the rules table for the candlepin version the rules
originated from. This is set on import, as well as during a manual
upload (to the current version of candlepin). Liquibase script will
set all existing rules entries to version 0.0.0 so we know they'll
always show up as older than any version of candlepin with this code.