Fix a bug with distributor manifest conflicts.
If you attempted to import and got a distributor conflict, the next
attempt would cause a MANIFEST_SAME conflict. Candlepin was not
correctly rolling back a change where we record the last import date and
the system mistakenly would think it had already imported this manifest.

Caused by @Transactional on a private method, which does not do
anything, so rollbacks were not kicking in correctly.