Performance improvements around anonymous entry points.
This should not be a huge benefit, but it allows for methods to mark themselves as anonymous. When
this is done, no principal is created. This is done primarily for the status and root controllers.

Other changes include:

1) Adding an authrnficationRequied method to the principal This method allwos
the security interceptor code to not use reflection to look for annotations unless it is necessary.

2) Only generate the links in the root controller once. This is really not related ot this patch, but it pissed me
off.