Improve performance of subscribe action in standalone mode.
For large subscriptions which provided bonus pools, there is a linear performance
degredation as the number of subscriptions increases. This is due to iterating over
all bonus pools in memory to find those which are overconsumed. This patch pushes the
calculation of what pools are over consumed ot the database, so that we are only iterating
over those which need to change in memory.