Add script to clean up the content ID changed breakage.
Content IDs were changed in upstream product service at some point (not
expected to happen again but the new IDs are permanent), leading to a
serious issue where manifests could no longer be imported as they
contained content with the same label as something already in db, but a
new ID.

The ID was the key we used for all product relationships and environment
promotions, across all orgs. Having it change is extremely difficult to
fix, we were unable to come up with an automated way to handle on the
fly in code and had to resort to a script to directly hit the affected
databases.

This will need to be run in any deployment where this issue arises. Once
run it should be fixed for good as the db will have the new IDs.