874785: null pointer while migrating owner
The MigrateOwnerJob contains a series of log statements that were
not checking for null before calling toString(). This patch adds
null checking to the parts that need it (i.e. won't cause catastrophic
failure later on).

For example, I don't check the consumer for being null, that's just
going to fail miserably anyway so no point in the log statements.