879022: Add message for too many content sets when V3 is disabled.
When cert V3 is disabled on the server and a client who supports
V3 certs attempts to consume an entitlement with too many content
sets, the error message would say that the client should be updated.

This is not the case if the client supports V3.0 certs... It is
because of the server, not the client.