Add a setter for ID on ProductContent, and a buildfile fix
ProductContent did not contain a setter for ID, which made it impossible to
serialize the object and then deserialize again. A dummy setter has been added,
to allow for deserialization from json.

Additionally, the custom candlepin exception classes were missing from the
certgen jar. I am not sure how this was missed, but it needs to be in there,
since we need to handle IseExceptions from library calls to candlepin.