884694: Fix import of manifests into older candlepin.
This logger argument was introduced in 0.7.9, but we cannot make this
kind of change without breaking import into all previous candlepin's
with the framework we have today. (they do not make the log argument
available to the javascript, so any attempt to run it will fail)

This issue is a little hard to hit in that it only occurs after the bad
rules are imported, and another import is attempted which contains new
subscriptions. (the original import succeeds because it uses the old
rules for the duration of the request)

Have to remove the logger for now.

Changes originated in commit: 9c9e3d62