858759: Do not remove guest virt ents when no host required
When guest IDs are updated in hosted mode, do not remove
the guest's virt entitlements if requires_host is _not_ set
on the pool. This generally means that candlepin is running
in hosted mode and uses a 'flex guest' model. Regardless of
the guest's host, the guest can rightfully have the
entitlement.

Found a test covering this scenario that was falsly passing,
hence the bug. It is now fixed up and properly tests this
BZ fix.