885857: Fix missing dependencies in non-cpdeps builds.
We recently removed the bundled jars in the candlepin rpm, which exposed
an issue with the logic around requires, simple syntax error was hiding
the deps.

Once this was fixed we found that serveral Requires were actually
missing, so these are added as well.

Allow tomcat6 to read symlinked jar files

840086: no longer bundle dependencies in war file
837655: no longer bundle dependencies in war file