Fix attempts to delete null cron triggers.
We do not fully understand why this can be but it looks like the old
code accommodated it, and we've run into problems with it being removed
already. Re-adding this check and skipping attempts to delete.

We will investigate what exactly it means for a cron job detail to exist
but the trigger doesn't.