Export both new and old rules.
We need to continue to include rules/default-rules.js for older
candlepin servers to be able to import the newer manifests.

Because Candlepin was set up to assume there was a single file in rules/
sub-dir and just import it, we have to move the new rules elsewhere as
there is no guarantee which order the files will return in listFiles().