906438: prevent saving status result from killing job
There was an exception during the job run that we tried to log in the
results column of the database. The column is only 255 characters
but the exception was clearly longer. So we caused yet another exception
by simply trying to write the result to the database.

I know log the exception to the log file then truncate the message
to 255 for storing. I also added more logging to the PinsetterJobListener
to hopefully get more information in case of errors.