908483: Add consumer types for katello.
The liquidbase scripts take in a community property. If this is set to True, then the types
Headpin and Katello are loaded. If this is set to False, then the types SAM and Katello are loaded.
The types share the same id, so the changelogs are mutually exclusive.

The cpdb and deploy tools are updated to pass in the new property. For Katello to take advantage of this, the katello configure tools will need to be updated.