Introduce candlepin software collection.
Due to a requirement on Rhino 1.7R3 and rules around shipping newer
versions of base OS packages, we have introduced a candlepin-scl
software collection. This in turn can be used by packages which we
branch and build into that scl, which will not conflict with other
versions of the same package.

This required changes to support looking for jars to symlink into our
webapp in alternate locations for any dependency we move into our SCL.

Add dep on ant-nodeps for the ant symlink task.