findbugs: CriteriaRules.log isn't final but should be
This static field public but not final, and could be changed by malicious code
or by accident from another package. The field could be made final to avoid this
vulnerability.

NOTE: I made it private instead of protected like the other uses of Logger.
This avoids the need for it being final and needing to be all caps: LOG
because of checkstyle.