New api for updating consumer lastCheckin time.
As part of the spacewalk integration work, we needed a way to get the last
checkin time of a spacewalk system into candlepin. However, candlepin only
allowed updating the last checkin to the current time, and only via the auth
interceptor.

The new updateLastCheckin() method allows for arbitrary last checkin times to
be set. It currently only needs to be called via admin; consumers will not be
using this directly.