fix random test failure
At times the test will fail because the date used to set the lastcheckin
is precisely the same as the date generated by the updateLastcheckin
method. This only affects the test because we're creating an arbitrary
situation to see if the method is actually changing the value. In
a production setting the setting and updating would not be this close.

I've added a 5ms sleep to give it just enough of a window to consistently
pass.