findbugs: Possible null pointer dereference of user
NP: Possible null pointer dereference of user in
org.candlepin.resource.ConsumerResource.verifyPersonConsumer(Consumer,
ConsumerType, Owner, String) on exception path

A reference value which is null on some exception control path is dereferenced
here.  This may lead to a NullPointerException when the code is executed.  Note
that because FindBugs currently does not prune infeasible exception paths, this
may be a false warning.

Also note that FindBugs considers the default case of a switch statement to be
an exception path, since the default case is often infeasible.