A developer script used to attach idcert to upstream consumer
When uploading an older manifest without an idcert embedded,
the identity cert will not be attached to the upstream consumer.
This script allows you to pass in a consumer.json to extract
the idcert from it and attach it to the upstream consumer that
matches the given uuid parameter value.