949684: Update the contract information on pools when subcriptions changed.
The code was not looking for these attributes to change. Code was added to copy them over only if htey have changed. Any certificates which have been generated will not be re-generated by this fix, as it seems like overkill to regen them. That could be added later if necessary.