Virt-limit sub-pool quantity should no longer use entitlement quantity.
When a physical system binds to a virt-limited pool, the sub-pool
created for guests of the host should have:

  quantity = virt_limit

Not what we do today:

  quantity = virt_limit * entitlement quantity

There was a unit test covering changes to the virt limit which was using
an entitlement quantity 1, which hides the issue. As such I changed that
test to use an entitlement quantity > 1, and make sure we have coverage
for the scenario where this rolls out and existing pools need to have
their quantity reduced.

Spec tests adjusted to make sure changing the entitlement quantity also
has no bearing on the sub-pool quantity.