Better method of detecting when to apply instance multiplier.
Previously we were using config for hosted/standalone, which means no
multiplier would be applied in standalone even if the subscription was
being created there. (custom products, dev testing, etc)

Instead we want to know if the subscription has been exported, if so we
know we do not need to re-apply multipliers.