ensure no test uses /etc/candlepin/candlepin.conf
In a previous commit 45b2aaec0 we changed CandlepinCommonTestConfig not
to use candlepin.conf. But we didn't do enough, there were many tests
that simply called the default constructor of Config.

I changed the tests to either use CandlepinCommonTestConfig() or
to create a custom Config passing in their own Map with data used
by the specific tests where they didn't need a full on Config.

I also changed the EventSinkImpl to let guice pass in its Config
instead of it constructing one (causing it to reload the config)
simply to get a value. It is better for it to use the existing
system's Config object.