heal entire org
Added a new API that will trigger the healing of the all the Consumers in an
Owner. The API returns a list of JobStatuses which the caller can use to check
the status of the of the jobs.

Added coded to handle scheduling multiple jobs based on the return of an array
of JobDetails.