965310: Fix broken import of identity cert.
Import was incorrectly trying to re-use a link to the original identity
certificate and it's serial. This worked fine if that existed in the
same db, as is often the case in our spec tests.

It doesn't work at all if you try the manifest on another server, or
wipe the db and then try to import the manifest.

Instead we will stay consistent with how EntitlementImporter gets around
this problem by creating a fake cert serial and making sure to create a
fresh copy of the identity certificate.

Spec tests modified to expose this by deleting the consumer we exported
from. These will fail on master, and pass with this fix.