963535: Fix instance quantity increment of 2 on virt guests.
We were returning the instance multiplier as the increment all the time,
which is incorrect for guests.

Further this is a rules style decision, so should be handled there.
Refactored QuantityRules to return both values we need with just the one
call. This unfortunately means a major rules API bump as there's no way
to do it without breaking compatability, or making a separate JS
invocation. However the previous major rules version is not yet released
so no major harm done.