remove cast to int
PoolRules had a FIXME comment where a long was being cast to an int
only to satisfy the call to Integer.toString(). If we simply wanted
a String of the value then String.valueOf() removes the need for
the cast, it will also keep the precision (not that we were
in danger of hitting it).