914827: Do not insert new consumer types if they are already present.
The manifest files pull down the latest consumer types. Therefore, the
scripts which work on the install fail on an upgrade. This patch causes
the inserts to be marked as having run if the content is already there.

Since the two types were added at the same time, there is no need to
test for both values