Track owner key and owner displayname on deletedconsumers
Previously, only the owner's PK was tracked in deleted consumer. This is not
helpful in the event that the consumer's owner was deleted, since there's no
way to look up what the rest of the owner info was, even with the PK.

Instead, track the owner key and displayname in addition to the primary key.