989698: Attempted fix for hornetq journal errors.
Unable to reproduce original bug with current config, however was able
to reproduce by removing explicit setting for journal size, and then
bumping the min-large message size up beyond the default value and
sending a huge message.

I believe it was incorrect to use the same value both for journal size,
and the min large message size, but I am not sure why the problem won't
happen locally. The large message size should always be smaller than the
journal buffer size. (equal is probably not correct either)

Attempted fix is to remove explicit setting of the journal size
and using the default ~500kb (which is much larger than we were setting)
Then bumping the min-large msg size to 100kb.