996925 - Exception while deleting manifest
The error is caused by the refresh pools. If a source pool is deleted before
the bonus pool, the deletion of the bonus pool causes the error in the trace.
In this scenario, the bonus pools will now not be deleted explicitly.

Also cleaned up the annotations as both javax.persistence.CascadeType.ALL and
hibernate.CascadeType.ALL were used on the same relationship. This has been shown
to cause conflicts. Removed the hibernate.CascadeType.MERGE because it is redundant.