994853: Fix installed product date range.
Bad assumption in the date range code, we were sorting on start date and
then assuming if the next entitlement in the list wasn't valid 1 second
after the last one ended, we had found a gap in coverage. This is not
accounting for multiple entitlements with the same end date, or an
entitlement with a later start date but an earlier end date.